requirements.txt
alembic==1.6.5
click==8.0.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
# itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.3.24
Werkzeug==2.0.1
astroid==2.2.5
ecdsa==0.13.2
future==0.17.1
isort==4.3.18
lazy-object-proxy==1.4.0
mccabe==0.6.1
pycryptodome==3.3.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
# typed-ast==1.4.2
wrapt==1.11.1